refactor: update Admonition to use config provider
