+++ docs/app/react/accordion/dev.mdx
@@ -2,6 +2,7 @@
 npm: '@cerberus-design/react'
 source: 'components/Accordion.tsx'
 recipe: 'slots/accordion.ts'
+ark: 'accordion'
 ---
 
 import CodePreview from '@/app/components/CodePreview'
