+++ docs/app/react/tabs/dev.mdx
@@ -2,6 +2,7 @@
 npm: '@cerberus-design/react'
 source: 'context/tabs.tsx'
 recipe: 'slots/tabs.ts'
+ark: 'tabs'
 ---
 
 import CodePreview from '@/app/components/CodePreview'
