+++ tests/react/components/admonition.test.tsx
@@ -4,22 +4,27 @@ import {
   Admonition,
   AdmonitionDescription,
   AdmonitionHeading,
+  CerberusProvider,
 } from '@cerberus-design/react'
-import { setupStrictMode } from '@/utils'
+import { makeConfig, setupStrictMode } from '@/utils'
 
 describe('Admonition', () => {
   setupStrictMode()
   afterEach(cleanup)
 
+  const config = makeConfig()
+
   test('should render a admonition element', () => {
     render(
-      <Admonition palette="page">
-        <AdmonitionHeading palette="page">When to use</AdmonitionHeading>
-        <AdmonitionDescription palette="page">
-          Use this component when you want to display a page-level message.{' '}
-          <a href="#">Learn more</a>
-        </AdmonitionDescription>
-      </Admonition>,
+      <CerberusProvider config={config}>
+        <Admonition palette="page">
+          <AdmonitionHeading palette="page">When to use</AdmonitionHeading>
+          <AdmonitionDescription palette="page">
+            Use this component when you want to display a page-level message.{' '}
+            <a href="#">Learn more</a>
+          </AdmonitionDescription>
+        </Admonition>
+      </CerberusProvider>,
     )
     expect(screen.getByText(/when to use/i)).toBeTruthy()
     expect(
@@ -31,13 +36,16 @@ describe('Admonition', () => {
 
   test('should render a admonition element with a page palette', () => {
     render(
-      <Admonition palette="page">
-        <AdmonitionHeading palette="page">When to use</AdmonitionHeading>
-        <AdmonitionDescription palette="page">
-          Use this component when you want to display a page-level message.{' '}
-          <a href="#">Learn more</a>
-        </AdmonitionDescription>
-      </Admonition>,
+      <CerberusProvider config={config}>
+        <Admonition palette="page">
+          <AdmonitionHeading palette="page">When to use</AdmonitionHeading>
+          <AdmonitionDescription palette="page">
+            Use this component when you want to display a page-level message.{' '}
+            <a href="#">Learn more</a>
+          </AdmonitionDescription>
+        </Admonition>
+        ,
+      </CerberusProvider>,
     )
     expect(
       screen
@@ -60,13 +68,15 @@ describe('Admonition', () => {
 
   test('should render a admonition element with a info palette', () => {
     render(
-      <Admonition palette="info">
-        <AdmonitionHeading palette="info">When to use</AdmonitionHeading>
-        <AdmonitionDescription palette="info">
-          Use this component when you want to display a page-level message.{' '}
-          <a href="#">Learn more</a>
-        </AdmonitionDescription>
-      </Admonition>,
+      <CerberusProvider config={config}>
+        <Admonition palette="info">
+          <AdmonitionHeading palette="info">When to use</AdmonitionHeading>
+          <AdmonitionDescription palette="info">
+            Use this component when you want to display a page-level message.{' '}
+            <a href="#">Learn more</a>
+          </AdmonitionDescription>
+        </Admonition>
+      </CerberusProvider>,
     )
     expect(
       screen
@@ -89,13 +99,15 @@ describe('Admonition', () => {
 
   test('should render a admonition element with a success palette', () => {
     render(
-      <Admonition palette="success">
-        <AdmonitionHeading palette="success">When to use</AdmonitionHeading>
-        <AdmonitionDescription palette="success">
-          Use this component when you want to display a page-level message.{' '}
-          <a href="#">Learn more</a>
-        </AdmonitionDescription>
-      </Admonition>,
+      <CerberusProvider config={config}>
+        <Admonition palette="success">
+          <AdmonitionHeading palette="success">When to use</AdmonitionHeading>
+          <AdmonitionDescription palette="success">
+            Use this component when you want to display a page-level message.{' '}
+            <a href="#">Learn more</a>
+          </AdmonitionDescription>
+        </Admonition>
+      </CerberusProvider>,
     )
     expect(
       screen
@@ -120,13 +132,15 @@ describe('Admonition', () => {
 
   test('should render a admonition element with a warning palette', () => {
     render(
-      <Admonition palette="warning">
-        <AdmonitionHeading palette="warning">When to use</AdmonitionHeading>
-        <AdmonitionDescription palette="warning">
-          Use this component when you want to display a page-level message.{' '}
-          <a href="#">Learn more</a>
-        </AdmonitionDescription>
-      </Admonition>,
+      <CerberusProvider config={config}>
+        <Admonition palette="warning">
+          <AdmonitionHeading palette="warning">When to use</AdmonitionHeading>
+          <AdmonitionDescription palette="warning">
+            Use this component when you want to display a page-level message.{' '}
+            <a href="#">Learn more</a>
+          </AdmonitionDescription>
+        </Admonition>
+      </CerberusProvider>,
     )
     expect(
       screen
@@ -151,13 +165,15 @@ describe('Admonition', () => {
 
   test('should render a admonition element with a danger palette', () => {
     render(
-      <Admonition palette="danger">
-        <AdmonitionHeading palette="danger">When to use</AdmonitionHeading>
-        <AdmonitionDescription palette="danger">
-          Use this component when you want to display a page-level message.{' '}
-          <a href="#">Learn more</a>
-        </AdmonitionDescription>
-      </Admonition>,
+      <CerberusProvider config={config}>
+        <Admonition palette="danger">
+          <AdmonitionHeading palette="danger">When to use</AdmonitionHeading>
+          <AdmonitionDescription palette="danger">
+            Use this component when you want to display a page-level message.{' '}
+            <a href="#">Learn more</a>
+          </AdmonitionDescription>
+        </Admonition>
+      </CerberusProvider>,
     )
     expect(
       screen
@@ -180,13 +196,15 @@ describe('Admonition', () => {
 
   test('should allow to render a custom icon', () => {
     render(
-      <Admonition icon={<span>🚀</span>}>
-        <AdmonitionHeading>When to use</AdmonitionHeading>
-        <AdmonitionDescription>
-          Use this component when you want to display a page-level message.{' '}
-          <a href="#">Learn more</a>
-        </AdmonitionDescription>
-      </Admonition>,
+      <CerberusProvider config={config}>
+        <Admonition icon={<span>🚀</span>}>
+          <AdmonitionHeading>When to use</AdmonitionHeading>
+          <AdmonitionDescription>
+            Use this component when you want to display a page-level message.{' '}
+            <a href="#">Learn more</a>
+          </AdmonitionDescription>
+        </Admonition>
+      </CerberusProvider>,
     )
     expect(screen.getByText(/🚀/)).toBeTruthy()
   })
